Most basic `gapless` function prototype
A simple API to create NextJS functions that run on your server, but can be run from the client context exactly as if they ran on the server. Seamlessly, with no extra work.

I'm thrilled with how well this has worked. Time to clean it up, optimize it, add better tooling and APIs on top of it, and really draw out its mega potential to shake up NextJS